+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # Changelog
 
+## [1.2.1] - 2022-02-xx
+
+### Minor changes
+
+- Context menu disabled on some pages that the extension is not allowed to read.
+
+
 ## [1.2.0] - 2022-02-20
 
 ### Major changes
 
-- AI user request is formatted with <pre> tag
+- AI user request is formatted with `<pre>` tag.
 - Decreased the number of unnecessary new lines after paragraphs.
 
 
@@ -13,7 +20,7 @@
 ### Major changes
 
 - Addon is now able to copy the **llama.cpp** AI conversations.
-- Added CS, DE, FR, IT translations
+- Added CS, DE, FR, IT translations.
 
 
 ## [1.0.0] - 2022-02-15